feat: rename pkg to internal
