+++ internal/daemon/subscribe.go
@@ -10,7 +10,7 @@ import (
 	"github.com/go-kratos/kratos/pkg/log"
 	"github.com/gorilla/websocket"
 	"github.com/itering/subscan/internal/substrate/rpc"
-	"github.com/itering/subscan/pkg/recws"
+	"github.com/itering/subscan/internal/pkg/recws"
 	"github.com/itering/subscan/internal/util"
 )
